MAINT: replace `Dish.ingrs` by a class
Each value in `Dish.ingrs` should contain information of weight, so
class `Ingredient` (the original one is renamed to `IngredientSheet` to
avoid confusion, so does `Menu` and `Seasoning`) is created for this.
Besides, `Dish.ingrs` is forced to be initialized with a list of
`Ingredient`. (If there is not existing data can be queried from
database, it will be initailized with a nameless `Ingredient` object)